requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
cssselect2==0.7.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
fonttools==4.45.1
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==3.6
MarkupSafe==2.1.3
pdfkit==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tinycss2==1.2.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
zopfli==0.2.3
